Refactor ChainIndicator and handleChangeNetwork
