+++ src/helpers/ChainHelper.ts
@@ -1,0 +1,36 @@
+// React + Web3 Essentials
+import { ethers } from 'ethers';
+
+// Internal Components
+import { PolygonNetworks } from 'helpers/UtilityHelper';
+import { appConfig } from 'config';
+
+// handles network change request
+export const handleChangeNetwork: (
+  chainId: number,
+  provider: ethers.providers.ExternalProvider
+) => Promise<void> = async (chainId, provider) => {
+  const chainIds = appConfig?.allowedNetworks;
+  if (chainIds.includes(chainId)) {
+    try {
+      await provider.request({
+        method: 'wallet_switchEthereumChain',
+        params: [{ chainId: ethers.utils.hexValue(chainId) }],
+      });
+    } catch (switchError) {
+      // This error code indicates that the chain has not been added to MetaMask.
+      if (switchError.code === 4902) {
+        const networkDetails = { ...PolygonNetworks[chainId === 137 ? 'POLYGON_MAINNET' : 'MUMBAI_TESTNET'] };
+        try {
+          await provider.request({
+            method: 'wallet_addEthereumChain',
+            params: [networkDetails],
+          });
+        } catch (addError) {
+          console.error(addError);
+        }
+      }
+      console.log(switchError);
+    }
+  }
+};
